requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.4
django-import-export==2.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
psycopg2-binary==2.8.6
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
